pyproject.toml
[tool.poetry]
name = "python-drf"
version = "0.1.0"
description = ""
authors = ["OlegMatus <129312840+OlegMatus@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pillow = "^11.1.0"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.7"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.4.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"